+++ src/main/java/com/github/mafelp/minecraft/listeners/CommandListener.java
@@ -1,4 +1,4 @@
-package com.github.mafelp.minecraft;
+package com.github.mafelp.minecraft.listeners;
 
 import com.github.mafelp.discord.DiscordMessageBroadcast;
 import com.github.mafelp.utils.Logging;
