+++ src/main/java/com/github/mafelp/minecraft/listeners/MinecraftChatListener.java
@@ -1,4 +1,4 @@
-package com.github.mafelp.minecraft;
+package com.github.mafelp.minecraft.listeners;
 
 import com.github.mafelp.accounts.Account;
 import com.github.mafelp.discord.ChannelAdmin;
